+++ src/constraints/generator_constraints.jl
@@ -36,7 +36,6 @@ function add_binGenIsOnInTS_constraints(m,p)
 	end 
 end
 
-
 function add_gen_can_run_constraints(m,p)
 	if p.s.generator.only_runs_during_grid_outage
 		for ts in p.time_steps_with_grid, t in p.techs.gen
@@ -58,6 +57,38 @@ function add_gen_rated_prod_constraint(m, p)
 	)
 end
 
+"""
+    add_generator_hourly_om_charges(m, p)
+
+- add decision variable "dvOMByHourBySizeGen" for the hourly Generator operations and maintenance costs
+- add the cost to TotalPerUnitHourOMCosts
+"""
+function add_generator_hourly_om_charges(m, p)
+    dv = "dvOMByHourBySizeGen"
+    m[Symbol(dv)] = @variable(m, [p.techs.gen, p.time_steps], base_name=dv, lower_bound=0)
+
+    #Constraint Generator-hourly-om-a: om per hour, per time step >= per_unit_size_cost * size for when on, >= zero when off
+	@constraint(m, GeneratorHourlyOMBySizeA[t in p.techs.gen, ts in p.time_steps],
+        p.s.generator.om_cost_per_hr_per_kw_rated * m[Symbol("dvSize")][t] -
+        (p.s.generator.existing_kw + p.s.generator.max_kw) * p.s.generator.om_cost_per_hr_per_kw_rated * (1-m[Symbol("binGenIsOnInTS")][t,ts])
+            <= m[Symbol("dvOMByHourBySizeGen")][t, ts]
+    )
+	#Constraint Generator-hourly-om-b: om per hour, per time step <= per_unit_size_cost * size for each hour
+	@constraint(m, GeneratorHourlyOMBySizeB[t in p.techs.gen, ts in p.time_steps],
+        p.s.generator.om_cost_per_hr_per_kw_rated * m[Symbol("dvSize")][t]
+            >= m[Symbol("dvOMByHourBySizeGen")][t, ts]
+    )
+	#Constraint Generator-hourly-om-c: om per hour, per time step <= zero when off, <= per_unit_size_cost*max_size
+	@constraint(m, GeneratorHourlyOMBySizeC[t in p.techs.gen, ts in p.time_steps],
+        (p.s.generator.existing_kw + p.s.generator.max_kw) * p.s.generator.om_cost_per_hr_per_kw_rated * m[Symbol("binGenIsOnInTS")][t,ts]
+            >= m[Symbol("dvOMByHourBySizeGen")][t, ts]
+    )
+    
+    m[:TotalHourlyGenOMCosts] = @expression(m, p.third_party_factor * p.pwf_om * 
+    	sum(m[Symbol(dv)][t, ts] * p.hours_per_time_step for t in p.techs.gen, ts in p.time_steps))
+    nothing
+end
+
 
 """
     add_gen_constraints(m, p)
@@ -69,6 +100,11 @@ function add_gen_constraints(m, p)
     add_binGenIsOnInTS_constraints(m,p)
     add_gen_can_run_constraints(m,p)
     add_gen_rated_prod_constraint(m,p)
+
+	m[:TotalHourlyGenOMCosts] = 0
+	if p.s.generator.om_cost_per_hr_per_kw_rated > 1.0E-7
+        add_generator_hourly_om_charges(m, p)
+    end
 
     m[:TotalGenPerUnitProdOMCosts] = @expression(m, p.third_party_factor * p.pwf_om *
         sum(p.s.generator.om_cost_per_kwh * p.hours_per_time_step *